% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_dictionaries.R
\docType{data}
\name{census_2011_dictionary}
\alias{census_2011_dictionary}
\title{ABS census 2011 data dictionary}
\format{
A list of data.frames:
\describe{
\item{dwelling}{data description of the dwelling table}
\item{family}{data description of the family table}
\item{person}{data description of the person table}
\item{areaenum}{data description of the place of
enumeration fields}
\item{statistical_areas}{data description of the statistical area standard}
}
}
\usage{
census_2011_dictionary
}
\description{
ABS census 2011 data dictionary
}
\keyword{datasets}
