% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fitting_control.R
\name{create_fitting_control}
\alias{create_fitting_control}
\title{Create a control table.}
\usage{
create_fitting_control(control_table)
}
\arguments{
\item{control_table}{a \code{data.table::data.table()}.}
}
\value{
a \code{data.table::data.table()} in the format that
requires by \code{mlfit::ml_problem()}.
}
\description{
Create a control table.
}
