% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fitting_sample.R
\name{create_fitting_sample}
\alias{create_fitting_sample}
\title{Create a reference sample}
\usage{
create_fitting_sample(microdata)
}
\arguments{
\item{microdata}{a \code{data.table::data.table()}.}
}
\value{
a \code{data.table::data.table()} in the format that
requires by \code{mlfit::ml_problem()}.
}
\description{
Create a reference sample.
}
