% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csf_geo.R
\name{csf_geo}
\alias{csf_geo}
\alias{csf_geo2016}
\alias{csf_geo2011}
\title{Create a census Samlple File's zone table}
\usage{
csf_geo2016()

csf_geo2011()
}
\description{
For creating a geographical hierachy table for \code{mlfit::ml_problem()}.
}
