% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_abs_shp.R
\name{load_abs_shp}
\alias{load_abs_shp}
\title{Load ABS Shapefile}
\usage{
load_abs_shp(url, crs = 4326)
}
\arguments{
\item{url}{a \emph{URL} to a shapefile on ABS website.}

\item{crs}{a numeric value that represents the spatial reference system
of the shapefile. The default is EPGS "4326" representing "WGS84".}
}
\value{
a \emph{sf} object.
}
\description{
This function loads a ABS shapefile from an URL. You may use this
\href{https://dbr.abs.gov.au/absmaps/index.html}{link} to help you select
an ABS SHP file you want to read.
}
\examples{
url <- "https://www.abs.gov.au/AUSSTATS/subscriber.nsf/log?openagent&1270055001_sa4_2016_aust_shape.zip&1270.0.55.001&Data\%20Cubes&C65BC89E549D1CA3CA257FED0013E074&0&July\%202016&12.07.2016&Latest"
load_abs_shp(url)
}
