% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph.R
\name{make_recode_cvista}
\alias{make_recode_cvista}
\title{A factory function to create \verb{recode_cvista_*()}}
\usage{
make_recode_cvista(.fn, .cols, .new_col, ...)
}
\arguments{
\item{.fn}{a function}

\item{.cols}{columns to use for conversion}

\item{.new_col}{name of the new column name}

\item{...}{dots}
}
\description{
A factory function to create \verb{recode_cvista_*()}
}
