% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph.R
\name{morph_cvista}
\alias{morph_cvista}
\title{Morph VISTA-population}
\usage{
morph_cvista(x, cvista_only = FALSE, extra_cols = NULL)
}
\arguments{
\item{x}{VISTA or Census population.}

\item{cvista_only}{a logical value. Only keep the recoded columns if \code{TRUE}.
Also, columns with 'cvista_' prefix will be renamed without the prefix.}

\item{extra_cols}{a character vector. Names of extra columns to be kept
after the recodings are done. These are exempted from \code{cvista_only}.
By default, ID columns will not be removed even when \code{cvista_only} is \code{TRUE}.}
}
\description{
This function helps to convert a Census population to a VISTA-like
population, which we refer to as CVISTA. The reason that we call this
a VISTA-like population is because many variables of VISTA population
doesn't exist or cannot be recreated using variables in Census. In some
cases, we need to create a new variable based on a set of variables from
both datasets, See the Notes section below about 'Derived variables'.
Hence, the final population will be more similar to VISTA than Census.
}
\note{
All variables with '_c' prefix mean that they are new variables, meaning they
could not be fully mapped to the variable of VISTA. For example, the
\code{relationship} variable of VISTA can only be partially recreated with
variables from CSF2016. See \code{recode_cvista_relationship()} for more details.

\strong{Derived variables}
employment type fields on both data sets do not map well to each other.
Hence, we need to create a new employment type variable which aligns better
to the Census population, since it provides a lower resolution.
}
