% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationSynthesiser-package.R
\docType{package}
\name{PopulationSynthesiser-package}
\alias{PopulationSynthesiser}
\alias{PopulationSynthesiser-package}
\title{PopulationSynthesiser: A Synthetic Population Generator for Australian Census Data}
\description{
| This packge contains implementations of population synthesiser approaches and utility functions for using Australian census data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AURIN-OFFICE/RCITI_PopulationSynthesiser}
}

}
\author{
\strong{Maintainer}: Amarin Siripanich \email{a.siripanich@unsw.edu.au} (\href{https://orcid.org/0000-0002-9140-4865}{ORCID})

Authors:
\itemize{
  \item Taha Rashidi \email{rashidi@unsw.edu.au} (\href{https://orcid.org/0000-0002-0673-5011}{ORCID})
}

Other contributors:
\itemize{
  \item Australian Urban Research Infrastructure Network (AURIN) [funder]
}

}
\keyword{internal}
