% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postfit_add_relationships.R
\name{postfit_add_relationships}
\alias{postfit_add_relationships}
\title{Add relationship fields to a Census synthetic population}
\usage{
postfit_add_relationships(x)
}
\arguments{
\item{x}{a data.table containing records of a Census synthetic population.}
}
\value{
a data.table
}
\description{
Add \code{partner_id}, \code{father_id}, and \code{mother_id} fields to the synthetic population
based on \code{rlhp} (relationship in household) and \code{sexp} (sex of person) fields.
}
