% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postfit_integerise_age5p.R
\name{postfit_integerise_age5p}
\alias{postfit_integerise_age5p}
\title{Turn 5 year age group into a singly year}
\usage{
postfit_integerise_age5p(x)
}
\arguments{
\item{x}{a data.table.}
}
\value{
a data.table with a new \code{age} field.
}
\description{
Turn 5 year age group into a singly year
}
