% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefit_control_retotal.R
\name{prefit_control_retotal}
\alias{prefit_control_retotal}
\title{Retotal a control table using a new grand total.}
\usage{
prefit_control_retotal(
  fitting_problem,
  individual_total = NULL,
  group_total = NULL,
  round = TRUE
)
}
\arguments{
\item{fitting_problem}{an \code{mlfit::ml_problem()} object.}

\item{individual_total}{a new grand total of the person control tables.}

\item{group_total}{a new grand total of the group control tables.}

\item{round}{a logical value whether to round the new totals. Default as \code{TRUE}.}
}
\value{
a new control table where the grand total equals to \code{total} where
the proportion of each cell remains the same as the original control table.
}
\description{
This is useful when you want to simply project a control
table for a future year where a new grand total is assumed
but the distribution of the control remains unchanged.
}
