% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefit_csf_agep_to_age5p.R
\name{prefit_csf_agep_to_age5p}
\alias{prefit_csf_agep_to_age5p}
\title{Recategorise Census Sample File's Age of Person to five-year age group}
\usage{
prefit_csf_agep_to_age5p(fitting_problem)
}
\arguments{
\item{fitting_problem}{an \code{mlfit::ml_problem()} object.}
}
\value{
an updated \code{mlfit::ml_problem()} object
}
\description{
Turn \code{agep} into \code{age5p} (five-year age groups).
}
