% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_microdata.R
\docType{data}
\name{read_census_microdata}
\alias{read_census_microdata}
\alias{csf_id_cols}
\title{Read ABS Census of Population and Housing Microdata into a tidy format}
\format{
An object of class \code{character} of length 3.
}
\usage{
read_census_microdata(
  person_file,
  family_file,
  dwelling_file,
  census_year = c("2011", "2016"),
  label = TRUE,
  select = NULL,
  include_overseas_visitors = FALSE
)

csf_id_cols
}
\arguments{
\item{person_file}{\code{character(1)} a path to a person microdata CSV file}

\item{family_file}{\code{character(1)} a path to a family microdata CSV file}

\item{dwelling_file}{\code{character(1)} path to a dwelling microdata CSV file}

\item{census_year}{\code{character(1)} denote the census year of the input
census microdata}

\item{label}{\code{logical(1)} should the microdata be labelled? default as \code{TRUE}}

\item{select}{\code{character()} A vector of column names to keep. However,
\code{abshid}, \code{abspid} and \code{absfid} are kept, regardless of what is in \code{select}.
By default, no columns will be discarded.}

\item{include_overseas_visitors}{\code{logical(1)}, default as \code{FALSE}.
If \code{TRUE} all overseas visitors, those that have "RLHP" equal to
"Overseas visitors", will be included.}
}
\value{
a \code{csf} object, which is also a data.table.
}
\description{
This function reads
\href{https://www.abs.gov.au/ausstats/abs@.nsf/Latestproducts/2037.0.30.001Main\%20Features32016?opendocument&tabname=Summary&prodno=2037.0.30.001&issue=2016&num=&view=}{ABS Population and Housing Microdata}
into a tidy format that allows you to create a fitting problem.
}
\note{
\code{csf_id_cols}: ID columns of CSF 2011 and 2016
}
\keyword{datasets}
