% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_tablebuilder.R
\name{read_census_tablebuilder}
\alias{read_census_tablebuilder}
\title{Read Census TableBuilder generated csv files}
\usage{
read_census_tablebuilder(x, remove_totals = TRUE, simplify_names = TRUE)
}
\arguments{
\item{x}{\code{character(1)} the directory to the file to be read.}

\item{remove_totals}{\code{logical(1)} Default as \code{TRUE}. If \code{TRUE}
all the rows that contains 'Total' will be removed. Any fields
that had the summation button selected on the Census TableBuilder
will contain rows with 'Total' as a category.}

\item{simplify_names}{\code{logical(1)} Default as \code{TRUE}. Only keep
the abbrevatiated names, if available, and convert everything to
names that comply with R's variable name restriction, except
everything will be in lowercase.}
}
\value{
a \code{data.table::data.table()}.
}
\description{
This function can read
\href{https://www.abs.gov.au/websitedbs/censushome.nsf/home/tablebuilder}{Census TableBuilder}
generated csv files without the users having to strip the
metadata from the files.
}
\examples{
ctb_testdata_dir <-
  system.file("testdata", "2016-aus-ur-agep.csv",
    package = "PopulationSynthesiser"
  )
read_census_tablebuilder(ctb_testdata_dir)
}
