% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph.R
\name{recode_cvista_age5}
\alias{recode_cvista_age5}
\alias{recode_cvista_age5.csf2016}
\alias{recode_cvista_age5.vista1218}
\title{Create CVISTA five-year age group (derived variable)}
\usage{
recode_cvista_age5(x, ...)

\method{recode_cvista_age5}{csf2016}(x, remove_cols = FALSE, ...)

\method{recode_cvista_age5}{vista1218}(x, remove_cols = FALSE, ...)
}
\arguments{
\item{x}{a \code{csf2016} data.table or a \code{vista1218} data.table.}

\item{...}{dots}

\item{remove_cols}{remove the recoding columns}
}
\description{
Create a five-year age group variable using
VISTA1218's \code{age} (age of person) and CSF2016's \code{agep}
(age of person)
}
