% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph.R
\name{recode_cvista_cars}
\alias{recode_cvista_cars}
\alias{recode_cvista_cars.csf2016}
\alias{recode_cvista_cars.vista1218}
\title{Create CVISTA cars (partial match)}
\usage{
recode_cvista_cars(x, ...)

\method{recode_cvista_cars}{csf2016}(x, remove_cols = FALSE, ...)

\method{recode_cvista_cars}{vista1218}(x, remove_cols = FALSE, ...)
}
\arguments{
\item{x}{a \code{csf2016} data.table or a \code{vista1218} data.table.}

\item{...}{dots}

\item{remove_cols}{remove the recoding columns}
}
\description{
VISTA1218's \code{CARS} variable or "Number of cars" can be
\strong{partially} recreated with CSF2016's \code{vehrd} (number of vehicles).
}
