% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph.R
\name{recode_cvista_emptype}
\alias{recode_cvista_emptype}
\alias{recode_cvista_emptype.csf2016}
\alias{recode_cvista_emptype.vista1218}
\title{Create CVISTA employment type}
\usage{
recode_cvista_emptype(x, ...)

\method{recode_cvista_emptype}{csf2016}(x, remove_cols = FALSE, ...)

\method{recode_cvista_emptype}{vista1218}(x, remove_cols = FALSE, ...)
}
\arguments{
\item{x}{a \code{csf2016} data.table or a \code{vista1218} data.table.}

\item{...}{dots}

\item{remove_cols}{remove the recoding columns}
}
\description{
VISTA1218's \code{EMPTYPE} variable or employment type can be recreated
with CSF2016's SIEMP (Status in Employment).
}
\note{
If SIEMP is 'Owner with not stated employee', it will be assumed that
they are 'Employer'.
}
