% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph.R
\name{recode_cvista_hhinc}
\alias{recode_cvista_hhinc}
\alias{recode_cvista_hhinc.csf2016}
\alias{recode_cvista_hhinc.vista1218}
\title{Create CVISTA hhinc (derived variable)}
\usage{
recode_cvista_hhinc(x, ...)

\method{recode_cvista_hhinc}{csf2016}(x, remove_cols = FALSE, ...)

\method{recode_cvista_hhinc}{vista1218}(x, remove_cols = FALSE, ...)
}
\arguments{
\item{x}{a \code{csf2016} data.table or a \code{vista1218} data.table.}

\item{...}{dots}

\item{remove_cols}{remove the recoding columns}
}
\description{
VISTA1218's \code{HHINC} variable or "Household income" can be
used to recreate with CSF2016's \code{rlhp} (Relatioship in household)
and \code{rpip} (Family/Household Reference Person Indicator).
}
\note{
"Negative income" in CSF2016 will be recoded as "Nil income" as
\code{hhinc} in VISTA1218 doesn't contain negative income.
}
