% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph.R
\name{recode_cvista_studying}
\alias{recode_cvista_studying}
\alias{recode_cvista_studying.csf2016}
\alias{recode_cvista_studying.vista1218}
\title{Create CVISTA studying}
\usage{
recode_cvista_studying(x, ...)

\method{recode_cvista_studying}{csf2016}(x, remove_cols = FALSE, ...)

\method{recode_cvista_studying}{vista1218}(x, remove_cols = FALSE, ...)
}
\arguments{
\item{x}{a \code{csf2016} data.table or a \code{vista1218} data.table.}

\item{...}{dots}

\item{remove_cols}{remove the recoding columns}
}
\description{
VISTA1218's \code{STUDYING} variable or "Student Status - level or none?" can be recreated
with CSF2016's \code{STUP} (Full/Part Time Student Status) and \code{TYPP}
(Type of Educational Institution Attending).
}
\note{
If SIEMP is 'Owner with not stated employee', it will be assumed that
they are 'Employer'.
}
