% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_workflow.R
\name{run_workflow}
\alias{run_workflow}
\title{Run the ATRC workflow}
\usage{
run_workflow(config_path = "/atrc_data/parameters.yaml")
}
\arguments{
\item{config_path}{path to an ATRC workflow config file.
See the expected format at \url{https://github.com/ATRC-AURIN/PopulationSynthesiser/blob/main/atrc_workflow/parameters.yaml}.}
}
\value{
a list of \code{mlfit}'s fitted problem and a \code{data.frame}
of the resulting synthetic population.
}
\description{
This workflow creates a synthetic population from a reference sample
and control data as specified in the input config file. The fitting and
replicating algorithms can also be selected in the config file.
}
