% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vista.R
\docType{data}
\name{vista1218web}
\alias{vista1218web}
\title{A merged between the persons table and the households table of VISTA
2012-2018, a public version.}
\format{
A data frame with 64516 rows and 71 variables:
Please see the official data glossary by downloding the data from source, linked below.
}
\source{
\url{https://transport.vic.gov.au/about/data-and-research/vista/vista-data-and-publications}
}
\usage{
vista1218web
}
\description{
A merged between the persons table and the households table of VISTA
2012-2018, a public version.
}
\keyword{datasets}
