% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atrc_workflow.R
\name{run_workflow}
\alias{run_workflow}
\title{Run the ATRC workflow}
\usage{
run_workflow(config_path = "/atrc_data/parameters.yaml")
}
\arguments{
\item{config_path}{path to an ATRC workflow config file.
See the Details section below for the expected format.}
}
\value{
a list of \code{mlfit}'s fitted problem and a \code{data.frame}
of the resulting synthetic population.
}
\description{
This workflow creates a synthetic population from a reference sample
and control data as specified in the input config file. The fitting and
replicating algorithms can also be selected in the config file.
}
\details{
Here is the expected format of a config file.

\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{inputs:
  REFERENCE_SAMPLE:
    type: path
    path: /atrc_data/inputs/reference_sample.csv
  PERSONS_CONTROL:
    type: multiple_paths
    path:
      - /atrc_data/inputs/persons-control1.csv
  HOUSEHOLDS_CONTROL:
    type: multiple_paths
    path:
      - /atrc_data/inputs/households-control1.csv
  GROUP_ID:
    type: value
    value: "hhid"
  INDIVIDUAL_ID:
    type: value
    value: "persid"
  COUNT:
    type: value
    value: "n"
  ML_FIT_ALGORITHM:
    type: value
    value: "ipu"
  ML_REPLICATE_ALGORITHM:
    type: value
    value: "trs"
outputs:
  SYNTHETIC_POPULATION:
    type: csv
    path: /atrc_data/outputs/synthetic_population.csv
}\if{html}{\out{</div>}}
}
