% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_workflow.R
\name{run_workflow}
\alias{run_workflow}
\title{Run the ATRC workflow}
\usage{
run_workflow(config_path = "/atrc_data/parameters.yaml")
}
\arguments{
\item{config_path}{path to an ATRC workflow config file.
See the Details section below for the expected format.}
}
\value{
a list of \code{mlfit}'s fitted problem and a \code{data.frame}
of the resulting synthetic population.
}
\description{
This workflow creates a synthetic population from a reference sample
and control data as specified in the input config file. The fitting and
replicating algorithms can also be selected in the config file.
}
\details{
Here is the expected format of a config file.

\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{inputs:
  REFERENCE_SAMPLE:
    description: The reference sample to use for the synthetic population.
    type: path
    path: atrc_data/inputs/reference_sample.csv
  PERSONS_CONTROL:
    description: "The person-level control files."
    type: multiple_paths
    path:
      - atrc_data/inputs/persons-control1.csv
      - atrc_data/inputs/persons-control2.csv
  HOUSEHOLDS_CONTROL:
    description: "The household-level control files."
    type: multiple_paths
    path:
      - atrc_data/inputs/households-control1.csv
      - atrc_data/inputs/households-control2.csv
  GROUP_ID:
    description: The group ID column of the reference sample.
    type: value
    value: "hhid"
  INDIVIDUAL_ID:
    description: The individual ID column of the reference sample.
    type: value
    value: "persid"
  COUNT:
    description: |
      Name of the control total column in control tables.
    type: value
    value: "n"
  ML_FIT_ALGORITHM:
    description: |
      The algorithm to use for fitting the synthetic population. 
      See the available algorithms at https://mlfit.github.io/mlfit/reference/ml_fit.html.
    type: value
    value: "ipu"
  ML_REPLICATE_ALGORITHM:
    description: |
      The algorithm to use for replicating the synthetic population.
      See the available algorithms at https://mlfit.github.io/mlfit/reference/ml_replicate.html.
    type: value
    value: "trs"
outputs:
  OUTPUT_DIRECTORY:
    description: The path where the output files will be saved.
    type: path
    path: atrc_data/outputs
  SYNTHETIC_POPULATION_FILENAME:
    description: The path where the synthetic population will be saved as.
    type: string
    value: synthetic_population.csv
}\if{html}{\out{</div>}}
}
